refactor: urlsafe base64 encoding in common
