+++ google/cloud/internal/openssl_util.h
@@ -37,32 +37,6 @@ namespace internal {
 StatusOr<std::vector<std::uint8_t>> SignUsingSha256(
     std::string const& str, std::string const& pem_contents);
 
-/**
- * Returns a Base64-encoded version of @p bytes. Using the URL- and
- * filesystem-safe alphabet, making these adjustments:
- * -  Replace '+' with '-'
- * -  Replace '/' with '_'
- * -  Right-trim '=' characters
- */
-template <typename Collection>
-inline std::string UrlsafeBase64Encode(Collection const& bytes) {
-  Base64Encoder encoder;
-  for (auto c : bytes) encoder.PushBack(c);
-  std::string b64str = std::move(encoder).FlushAndPad();
-  std::replace(b64str.begin(), b64str.end(), '+', '-');
-  std::replace(b64str.begin(), b64str.end(), '/', '_');
-  auto end_pos = b64str.find_last_not_of('=');
-  if (std::string::npos != end_pos) {
-    b64str.resize(end_pos + 1);
-  }
-  return b64str;
-}
-
-/**
- * Decodes a Url-safe Base64-encoded string.
- */
-StatusOr<std::vector<std::uint8_t>> UrlsafeBase64Decode(std::string const& str);
-
 }  // namespace internal
 GOOGLE_CLOUD_CPP_INLINE_NAMESPACE_END
 }  // namespace cloud
